let $storage->put return a "key"
which is a string that can be further used to retrieve things from the
storage. Currently the content of that key is just

    $thing->type . "/" . $thing->id

So it should look like

    blob/QL0AFWMIX8NRZTKeof9cXsvbvu8

The idea is that we want to make these key distributable, therefore it
is better that type of thing is part of the key.